embed: Fix getter naming snafu
The documented get_use_layout_size() getter method is still named
get_honor_stage_size() in the library API, even though it does not match
the setter, and it's undocumented and undeclared.

Let's preserve the ABI compatibility, and keep exporting the symbol, but
also let's define the appropriate getter, to avoid an undefined symbol
when developers use it.